Make framebuffers (incl. Window class) more secure
glViewport will be called automatically when a framebuffer is bound with
the correct size. What the correct size for a framebuffer is, isn't
obvious and OpenGL allows textures with different sizes to be attached
to a framebuffer. However, this is seldom used so to make it more secure
I added that constraint to the FrameBuffer class.